requirements.txt
asttokens
comm
contourpy==1.3.1
cycler==0.12.1
debugpy
decorator
exceptiongroup
executing
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
imageio==2.36.1
importlib_metadata
ipykernel
ipython
jedi
Jinja2==3.1.5
jupyter_client
jupyter_core
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline
mpmath==1.3.0
nest_asyncio
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging
parso
pexpect
pickleshare
pillow==11.1.0
platformdirs
prompt_toolkit
psutil
ptyprocess
pure_eval
Pygments
pyparsing==3.2.1
python-dateutil
pyzmq
scikit-image==0.25.0
scipy==1.15.1
six
stack_data
sympy==1.13.1
tifffile==2025.1.10
torch==2.5.1
torchvision==0.20.1
tornado
tqdm==4.67.1
traitlets
triton==3.1.0
typing_extensions
uv==0.5.20
wcwidth
zipp
